chore: create a enums folder and locate the payment_status at it.
