+++ vendaja/enums/payment_status.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class PaymentStatus(Enum):
+    AWAITING_PAYMENT='AWAITING_PAYMENT'
+    PROCESSING='PROCESSING'
+    COMPLETED='COMPLETED'
+    FAILED_PAYMENT='FAILED_PAYMENT'
